third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-text-decoration-expected.txt

Test to make sure text-decoration property returns values properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial value:
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Initial value (explicit):
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Value 'none':
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Value 'underline':
PASS computedStyle.textDecoration is 'underline solid rgb(0, 0, 0)'


Value 'overline':
PASS computedStyle.textDecoration is 'overline solid rgb(0, 0, 0)'


Value 'line-through':
PASS computedStyle.textDecoration is 'line-through solid rgb(0, 0, 0)'


Value 'underline overline line-through':
PASS computedStyle.textDecoration is 'underline overline line-through solid rgb(0, 0, 0)'


Value 'blink' (valid but ignored):
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Value 'overline overline' (invalid):
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Value 'underline blank' (invalid):
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Value '':
PASS computedStyle.textDecoration is 'none solid rgb(0, 0, 0)'


Parent gets 'underline' value:
PASS computedStyle.textDecoration is 'underline solid rgb(0, 0, 0)'


Ancestor should explicitly inherit value from parent when 'inherit' value is used:
PASS computedStyle.textDecoration is 'underline solid rgb(0, 0, 0)'


Ancestor should not implicitly inherit value from parent (i.e. when value is void):
FAIL computedStyle.textDecoration should be none. Was none solid rgb(0, 0, 0).


PASS successfullyParsed is true

TEST COMPLETE
